+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+#### [0.3.3] - 2015-08-27
+##### Fixed/Added
+- Fixed the project scaffold option to cache static assets
+- Added `factories` directory to scaffold
+
 #### [0.3.2] - 2015-08-10
 ##### Added/Removed/Fixed/Changed
 - Added default loaded JavaScript file (application.js)